+++ ReadingOutputFromFLYonPIC_v2/AtomicConfigNumberConversion.py
